refactoring to use fixture
