+++ test/e2e/run-all.sh
@@ -35,9 +35,3 @@ retry concurrently --kill-others \
   'yarn dapp' \
   'mocha test/e2e/metamask-ui.spec'
 
-retry concurrently --kill-others \
-  --names 'sendwithprivatedapp,e2e' \
-  --prefix '[{time}][{name}]' \
-  --success first \
-  'yarn sendwithprivatedapp' \
-  'mocha test/e2e/incremental-security.spec'
